% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{event_set}
\alias{event_set}
\title{Simulated years of losses (class)}
\usage{
event_set(ptr)
}
\arguments{
\item{ptr}{Internal.}
}
\value{
An \code{event_set} object with read-only properties \code{n_sims} and \code{seed}.
}
\description{
Created by \code{\link[=simulate_events]{simulate_events()}}.
}
\keyword{internal}
