% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{gam_model}
\alias{gam_model}
\title{Fitted GAM (class)}
\usage{
gam_model(ptr, terms, xlevels)
}
\arguments{
\item{ptr, terms, xlevels}{Internal.}
}
\value{
A \code{gam_model} object.
}
\description{
Returned by \code{\link[=gam_fit]{gam_fit()}}.
}
