% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{glm_model}
\alias{glm_model}
\title{Fitted GLM (class)}
\usage{
glm_model(ptr, terms, xlevels)
}
\arguments{
\item{ptr, terms, xlevels}{Internal.}
}
\value{
A \code{glm_model} object.
}
\description{
Returned by \code{\link[=glm_fit]{glm_fit()}}.
}
