% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{stacking_fit}
\alias{stacking_fit}
\title{Posterior stacking weights (class)}
\usage{
stacking_fit(ptr, models, covariates)
}
\arguments{
\item{ptr, models, covariates}{Internal.}
}
\value{
A \code{stacking_fit} object.
}
\description{
Returned by \code{\link[=bayes_stacking]{bayes_stacking()}} and \code{\link[=hierarchical_stacking]{hierarchical_stacking()}}.
}
